Extract types to separate submodule
