+++ chain/indexer/src/streamer/fetchers.rs
@@ -1,7 +1,5 @@
 //! Streamer watches the network and collects all the blocks and related chunks
 //! into one struct and pushes in in to the given queue
-use std::collections::HashMap;
-
 use actix::Addr;
 use futures::stream::StreamExt;
 use tracing::warn;
@@ -9,39 +7,9 @@ use tracing::warn;
 pub use near_primitives::hash::CryptoHash;
 pub use near_primitives::{types, views};
 
+use self::super::types::{ExecutionOutcomesWithReceipts, IndexerExecutionOutcomeWithReceipt};
 use super::INDEXER;
 use crate::streamer::errors::FailedToFetchData;
-
-pub type ExecutionOutcomesWithReceipts = HashMap<CryptoHash, IndexerExecutionOutcomeWithReceipt>;
-
-/// Resulting struct represents block with chunks
-#[derive(Debug)]
-pub struct StreamerMessage {
-    pub block: views::BlockView,
-    pub chunks: Vec<IndexerChunkView>,
-    pub receipt_execution_outcomes: ExecutionOutcomesWithReceipts,
-    pub state_changes: views::StateChangesKindsView,
-}
-
-#[derive(Debug)]
-pub struct IndexerChunkView {
-    pub author: types::AccountId,
-    pub header: views::ChunkHeaderView,
-    pub transactions: Vec<IndexerTransactionWithOutcome>,
-    pub receipts: Vec<views::ReceiptView>,
-}
-
-#[derive(Clone, Debug)]
-pub struct IndexerTransactionWithOutcome {
-    pub transaction: views::SignedTransactionView,
-    pub outcome: IndexerExecutionOutcomeWithReceipt,
-}
-
-#[derive(Clone, Debug)]
-pub struct IndexerExecutionOutcomeWithReceipt {
-    pub execution_outcome: views::ExecutionOutcomeWithIdView,
-    pub receipt: Option<views::ReceiptView>,
-}
 
 pub(crate) async fn fetch_status(
     client: &Addr<near_client::ClientActor>,
